eink_art_display/requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
inky==1.2.0
mypy==0.800
mypy-extensions==0.4.3
numpy==1.20.0
pathspec==0.8.1
Pillow==8.1.0
regex==2020.11.13
requests==2.25.1
smbus2==0.4.1
spidev==3.5
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3